+++ src/test/java/nextstep/member/util/MemberTestUtil.java
@@ -1,4 +1,4 @@
-package nextstep.member;
+package nextstep.member.util;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
