+++ src/test/java/nextstep/theme/ThemeE2ETest.java
@@ -1,6 +1,7 @@
 package nextstep.theme;
 
 import nextstep.theme.model.Theme;
+import nextstep.theme.util.ThemeTestUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
